Update main.py
Stage 2 of the project. Created an array of examples using information about planets in the "planets" folder. Used FewShotPromptTemplate to create a prompt template based and pass the examples to the LLM. Formatted the prompt so that the user can just pass the name of the planet and get all the information needed.